pyproject.toml
[tool.poetry]
name = "kestrel"
version = "0.0.1"
description = "Kestrel Project."
authors = ["JosephNK <nkw0608@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poe.tasks]
start = "uvicorn main:app --reload --port 8010"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.3.7"
langsmith = "^0.1.143"
langchain-community = "^0.3.7"
langchain-core = "^0.3.19"
langchain-experimental = "^0.3.3"
langchain-openai = "^0.2.9"
langchain-anthropic = "^0.3.0"
langchain-text-splitters = "^0.3.2"
langchain-upstage = "^0.3.0"
langchain-chroma = "^0.1.4"
langchain-cohere = "^0.3.1"
langchain-milvus = "^0.1.7"
langchain-google-genai = "^2.0.4"
langchain-huggingface = "^0.1.2"
langchainhub = "^0.1.21"
langgraph = "^0.2.52"
huggingface-hub = "^0.26.2"
openai = "^1.54.4"
deepl = "^1.20.0"
kiwipiepy = "^0.20.1"
konlpy = "^0.6.0"

# 데이터 처리 및 분석 관련 패키지
pandas = "^2.2.2"
rank-bm25 = "^0.2.2"

# 데이터베이스 및 캐시 관련 패키지
redis = "^5.0.3"
chromadb = "^0.4.24"

# PDF 및 파일 처리 관련 패키지
pymupdf = "^1.24.1"
pypdf = "^4.2.0"
pdfplumber = "^0.11.1"
pdfminer-six = "^20231228"
pymupdf4llm = "^0.0"

# 시각화 및 사용자 인터페이스 관련 패키지
matplotlib = "^3.8.4"
streamlit = "^1.32.2"
jupyter = "^1.0.0"
notebook = "^7.1.2"

# 딥러닝 및 머신러닝 관련 패키지
torch = "^2.2.2"
torchvision = "^0.17.2"
faiss-cpu = "^1.8.0"
open-clip-torch = "^2.24.0"

# 기타 유틸리티 및 필수 패키지
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
lxml = "^5.2.2"
pillow = "^10.3.0"
lark = "^1.1.9"
ragas = "^0.1.7"
unstructured = {version = "0.10.19", extras = ["all-docs"]}
arxiv = "^2.1.0"
tiktoken = "^0.7"
tenacity = "8.3.0"
pymilvus = "^2.4.4"
google-search-results = "^2.4.2"
protobuf = "3.20.3"
sqlalchemy = "^2.0.31"
llama-index-core = "^0.10.58"
llama-parse = "^0.4.9"
llama-index-readers-file = "^0.1.30"
flashrank = "^0.2.8"
docx2txt = "^0.8"
nest-asyncio = "^1.6.0"
rapidocr-onnxruntime = "^1.3.24"
seaborn = "^0.13.2"
grandalf = "^0.8"
rouge-score = "^0.1.2"
langchain-ollama = "^0.2.0"
mypy = "^1.11.2"
pinecone = "^5.3.1"
beautifulsoup4 = "^4.12.3"
uvicorn = "^0.32.1"
pyupbit = "^0.2.34"
poethepoet = "^0.31.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"